Overhaul the project
The main structure is the same as before, but everything has been
touched up. Most important to note, the project now comes with Gradle.
Along with that, the code got overhauled and unnecessary files were
dropped and the structure was touched up as well. All while maintaining
the same compatibility as before. *cough* As far as I am aware.