Move around variable definitions
If the plan is to support all the versions, then the version list
should be as easy to oversee and modify as possible. This commit
splits up the versions and their class/method/field names into
multiple lines - the first line being the unobfuscated definition
and the former ones being the respective versions, with 5 in each
line.