Dynamically load the hook details
The hook details are now loaded from the GitHub repository. The load
is not that high and if someone is going to be using YouTube, they
better have internet connectivity anyhow, so we can afford to reload
them each time the package (application) gets loaded. This means we
are completely sure users get the latest JSON when they update their
YouTube applications, restart the applications and restart their
entire devices. We could add some kind of caching later down the line
and the system can easily be adjusted to read in a file and pass
the parsed JSON object into the `hook` method, but I see no point
for us to do that now.

All rejoice. By the way, can we have a v2.0.0 release for this?