Preparing for 2.0 (json) release
* Trim trailing whitespace

That whitespace does not belong there, so this commit simply drops it.

* Add a modular JSON asset

We really need to redo the way we deal with YouTube application
updates. This commit pushes a new proposal for the data structure
of the version method details.

Each version has its own JSON array consisting of JSON objects
which specify the names of the class and the method which should
be hooked as well as the actions to do on the method. Currently
it seems reasonable to add support for field value setting before
and after the method as well as returning a constant value. Right
now the value types are boolean and string and can probably stay
that way for our intents and purposes. We do want to keep the
actions listed inside of the JSON file instead of being hardcoded
as speficic "action codes" in the codebase in case a new release
of YouTube changes around the structure (it's happened once in
the lifetime of this module already, so it can happen again).

Worth keeping in mind the contents of the new file are based on
the old `youtube_hooks.json` so if there were incorrect entries
in that file or I messed up the manual copy-paste restructure,
switching over to this file might mess things up. (Sorry if that
happens later down the line!)

* Add latest app release data to JSON file

The JSON was a bit out of date and missing some details. The method
specification now includes the signature (parameter types) specified
as the class names.

* Follow lowercase convention for field names

`parameter_types` instead of `parameterTypes` it is.