Support Xposed versions older than 82
Apparently, a helper method called `findClassIfExists` which I opted
to use during the hooking of methods is only available if you have
installed Xposed version 82 or newer. This is an issue since we have
a bunch of users who do not update their Xposed framework for whatever
reasons they might have as we should still support them.

This commit avoids the issue by switching to the plain `findClass` and
wrapping it in a `try-catch` block for the missing class exception.